Move options to a separate file
